+++ eggs/naaya-survey/Products/NaayaWidgets/widgets/MatrixWidget.py
@@ -17,8 +17,10 @@ class MatrixWidget(Widget):
 
     def __init__(self, id, lang=None, **kwargs):
         Widget.__init__(self, id, lang, **kwargs)
-        self.set_localproperty('choices', 'lines', lang, [])
-        self.set_localproperty('rows', 'lines', lang, [])
+        choices = kwargs.get('choices', [])
+        self.set_localproperty('choices', 'lines', lang, choices)
+        rows = kwargs.get('rows', [])
+        self.set_localproperty('rows', 'lines', lang, rows)
 
     def getChoices(self, anyLangNonEmpty=False):
         """ """